+++ persistent/Database/Persist/Quasi/Internal.hs
@@ -48,7 +48,7 @@ import Data.Text (Text)
 import qualified Data.Text as T
 import Database.Persist.Types
 import Text.Read (readEither)
-import Database.Persist.Names
+import Database.Persist.EntityDef.Internal
 
 data ParseState a = PSDone | PSFail String | PSSuccess a Text deriving Show