+++ persistent/Database/Persist/Types.hs
@@ -1,6 +1,7 @@
 module Database.Persist.Types
     ( module Database.Persist.Types.Base
     , module Database.Persist.Names
+    , module Database.Persist.EntityDef
     , SomePersistField (..)
     , Update (..)
     , BackendSpecificUpdate
@@ -14,6 +15,36 @@ module Database.Persist.Types
     ) where
 
 import Database.Persist.Names
-import Database.Persist.Types.Base
 import Database.Persist.Class.PersistField
 import Database.Persist.Class.PersistEntity
+import Database.Persist.EntityDef
+
+-- this module is a bit of a kitchen sink of types and concepts. the guts of
+-- persistent, just strewn across the table. in 2.13 let's get this cleaned up
+-- and a bit more tidy.
+import Database.Persist.Types.Base
+    ( FieldCascade(..)
+    , ForeignDef(..)
+    , CascadeAction(..)
+    , FieldDef(..)
+    , UniqueDef(..)
+    , FieldAttr(..)
+    , IsNullable(..)
+    , WhyNullable(..)
+    , ExtraLine
+    , FieldType(..)
+    , PersistException(..)
+    , ForeignFieldDef
+    , Attr
+    , CompositeDef(..)
+    , SqlType(..)
+    , ReferenceDef(..)
+    , noCascade
+    , parseFieldAttrs
+    , keyAndEntityFields
+    , PersistException(..)
+    , UpdateException(..)
+    , PersistValue(..)
+    , PersistFilter(..)
+    , PersistUpdate(..)
+    )